integrations/llamaindex/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lobster-academy-llamaindex"
version = "0.1.0"
description = "LlamaIndex integration for Lobster Academy - AI Agent behavior evidence"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
dependencies = [
    "llama-index-core>=0.10.0",
    "lobster-academy>=0.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio"]
